# mvcc: don't use pointer for storeTxnRead in storeTxnWrite