Saves an allocation when creating a storeTxnWrite.

Before:
```
BenchmarkStoreTxnPut-4            	  200000	      7728 ns/op	    1515 B/op	      16 allocs/op
BenchmarkWatchableStoreTxnPut-4   	  200000	      7934 ns/op	    1580 B/op	      18 allocs/op
```

After:
```
BenchmarkStoreTxnPut-4            	  200000	      7319 ns/op	    1498 B/op	      15 allocs/op
BenchmarkWatchableStoreTxnPut-4   	  200000	      7558 ns/op	    1565 B/op	      17 allocs/op
```